project: improve clean script

+++ scripts/clean.ts
@@ -10,8 +10,12 @@ export const clean = $command({
 		const a = `${p}/alepha`;
 		await run("rm -rf coverage");
 		await run(`rm -rf ${p}/**/dist ${p}/**/node_modules ${p}/**/coverage`);
-		const dirs = (await readdir(a)).filter((f) => f !== "src").map((f) => `${a}/${f}`).join(" ");
-		await run(`rm -rf ${dirs}`);
+		const dirs = (await readdir(a))
+			.filter((f) => f !== "src" && !f.includes(".") && f !== "LICENSE")
+			.map((f) => `${a}/${f}`).join(" ");
+		if (dirs.trim()) {
+			await run(`rm -rf ${dirs}`);
+		}
 		await run(`rm -rf ${a}/*.js ${a}/*.cjs ${a}/*.map ${a}/*.d.ts`);
 		await run("yarn");
 	},
